+++ hubdle-settings-gradle-plugin/main/kotlin/com/javiersc/hubdle/settings/HubdleSettingsExtension.kt
@@ -1,6 +1,7 @@
 package com.javiersc.hubdle.settings
 
 import com.javiersc.hubdle.settings.extensions.HubdleAutoIncludeExtension
+import com.javiersc.hubdle.settings.extensions.HubdleBuildScanExtension
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.model.ObjectFactory
@@ -14,5 +15,12 @@ public open class HubdleSettingsExtension @Inject constructor(objects: ObjectFac
     @HubdleSettingsDslMarker
     public fun autoInclude(action: Action<HubdleAutoIncludeExtension> = Action {}) {
         action.execute(autoInclude)
+    }
+
+    public val buildScan: HubdleBuildScanExtension = objects.newInstance()
+
+    @HubdleSettingsDslMarker
+    public fun buildScan(action: Action<HubdleBuildScanExtension> = Action {}) {
+        action.execute(buildScan)
     }
 }